refactor(secret-encryption): merge column rotation service into the handler
ColumnRotationSiteHandler was a thin config wrapper that delegated
every method to SecretEncryptionColumnRotationService. Merge the
service's countNonCurrentRows / rotateColumn / rotateRow logic into
the handler itself: the handler now owns its repository, config, and
SecretEncryptionService dependency, and there is no longer a separate
service to keep in sync.

Net effect: one less file, one less NestJS provider; the runner
injects SecretEncryptionService directly and passes it to each
ColumnRotationSiteHandler instance.
