+++ packages/twenty-server/src/database/commands/secret-encryption-rotation/secret-encryption-rotation.module.ts
@@ -3,7 +3,6 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { SensitiveConfigStorageRotationHandler } from 'src/database/commands/secret-encryption-rotation/handlers/sensitive-config-storage-rotation.handler';
 import { RotateSecretEncryptionCommand } from 'src/database/commands/secret-encryption-rotation/rotate-secret-encryption.command';
-import { SecretEncryptionColumnRotationService } from 'src/database/commands/secret-encryption-rotation/services/secret-encryption-column-rotation.service';
 import { SecretEncryptionRotationRunnerService } from 'src/database/commands/secret-encryption-rotation/services/secret-encryption-rotation-runner.service';
 import { ApplicationRegistrationVariableEntity } from 'src/engine/core-modules/application/application-registration-variable/application-registration-variable.entity';
 import { ApplicationVariableEntity } from 'src/engine/core-modules/application/application-variable/application-variable.entity';
@@ -28,7 +27,6 @@ import { ConnectedAccountEntity } from 'src/engine/metadata-modules/connected-ac
     ]),
   ],
   providers: [
-    SecretEncryptionColumnRotationService,
     SensitiveConfigStorageRotationHandler,
     SecretEncryptionRotationRunnerService,
     RotateSecretEncryptionCommand,
